fix(DiscountEvent): long을 int로 변경

+++ src/main/java/christmas/constant/DiscountEvent.java
@@ -14,7 +14,7 @@ public enum DiscountEvent {
     CHRISTMAS_D_DAY(
             "크리스마스 디데이 할인",
             today -> today.between(1, 25),
-            (today, orders) -> 1000L + (today.getToday() - 1) * 100L
+            (today, orders) -> 1000 + (today.getToday() - 1) * 100
     ),
     WEEKDAY(
             "평일 할인",
@@ -34,12 +34,12 @@ public enum DiscountEvent {
 
     private final String eventName;
     private final Function<Today, Boolean> validator;
-    private final BiFunction<Today, Orders, Long> discountCalculator;
+    private final BiFunction<Today, Orders, Integer> discountCalculator;
 
     DiscountEvent(
             String eventName,
             Function<Today, Boolean> validator,
-            BiFunction<Today, Orders, Long> discountCalculator
+            BiFunction<Today, Orders, Integer> discountCalculator
     ) {
         this.eventName = eventName;
         this.validator = validator;
